Code is executed in new app domain
New app domain is created, in which the test code is loaded and
executed. This allows the assembly to be unloaded and the new DLLs to be
cleaned up after use. This can be considered stable, although there is
some hard coding to my specific repository which I intend to refactor
out very soon.